Move `FigureFactory`/etc. into graph_objs package.
